+++ domain/request.go
@@ -30,5 +30,6 @@ type BroadcastEventRequest struct {
 
 type CreateUpdateConfigRequest struct {
 	entity.Config
-	Unsafe bool
+	Unsafe    bool
+	CreateNew bool
 }